pyproject.toml
[tool.poetry]
name = "uploadgram"
version = "0.3.3"
description ="Upload files to Telegram upto 4 GiB, from the Terminal"
authors = ["Shrimadhav U K <uploADGRam@shrimadhavUK.me>"]
homepage = "https://github.com/SpEcHiDe/UploadGram"
repository = "https://github.com/SpEcHiDe/UploadGram"
keywords = ["telegram-upload", "telegram", "upload", "video", "audio"]
license = "AGPLv3"
readme = "README.md"

classifiers = [
    "Development Status :: 4 - Beta",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development",
]

[tool.poetry.dependencies]
python = ">=3.7.7,<4"
python-dotenv = "0.10"
pyrotgfork = "2.1.33.8"
TgCrypto = "1.2.5"
hachoir = "3.1.1"
tqdm = "4.62.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.7.0"
black = "^20.8b1"
autopep8 = "^1.5.7"
autoflake = "^1.4"

[tool.isort]
profile = "black"
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
filter_files = true

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
uploadgram = "uploadgram.shell:main"
